Implement Maximum Binary Tree problem in leetcode-654.py
This commit fills out the 'construct_maximum_binary_tree' function in leetcode-654.py which was previously empty. The function uses recursion to build a maximum binary tree from a given list of integers, where each TreeNode has its left and right children defined as the maximum binary trees of the numbers to its left and right. The function 'max_range' is used to find the index of the maximum number in a range.