Add Node and Solution classes to leetcode-116.py
Created new classes Node and Solution to solve a problem on leetcode which is connecting all nodes within a binary tree. Currently, 'connect' method in Solution class is empty. This will be filled in future commits with logic to connect all nodes. Node class contains initializers for value, left, right and next pointers.