Implement Leetcode-1535 "Find the Winner of an Array Game"
Implemented the solution for Leetcode problem 1535, titled "Find the Winner of an Array Game". The python function getWinner takes an array and an integer 'k' as arguments and finds which number would win after 'k' comparisons in the array game. A number is deemed a winner if it surpasses consecutive 'k' numbers in the array.